removed original samples and replace them with new ones. modified new samples (reduce code duplication, add cmd line params and short description)
