+++ samples/directx/d3d9_interop.cpp
@@ -1,149 +1,297 @@
+#define WIN32_LEAN_AND_MEAN
 #include <windows.h>
 #include <d3d9.h>
+
+#include "opencv2/core.hpp"
+#include "opencv2/core/directx.hpp"
+#include "opencv2/core/ocl.hpp"
+#include "opencv2/imgproc.hpp"
+#include "opencv2/videoio.hpp"
+
+#include "d3dsample.hpp"
+
 #pragma comment (lib, "d3d9.lib")
 
-#define USE_D3D9
-#define WINDOW_NAME "OpenCV Direct3D 9 Sample"
 
-IDirect3D9 *pD3D = NULL;
-IDirect3DDevice9 *dev = NULL;
-IDirect3DSurface9 *pBackBuffer = NULL;
-IDirect3DSurface9 *pCPUWriteSurface = NULL; // required name
-IDirect3DSurface9 *pReadOnlySurface = NULL; // required name
-HANDLE readOnlySurfaceShared = 0; // required name
-IDirect3DSurface9 *pSurface = NULL; // required name
-HANDLE surfaceShared = 0; // required name
+using namespace std;
+using namespace cv;
 
-#include "d3d_base.inl.hpp"
-
-bool initDirect3D(void)
+class D3D9WinApp : public D3DSample
 {
-    if (NULL == (pD3D = Direct3DCreate9(D3D_SDK_VERSION)))
+public:
+    D3D9WinApp(int width, int height, std::string& window_name, cv::VideoCapture& cap) :
+        D3DSample(width, height, window_name, cap) {}
+
+    ~D3D9WinApp() {}
+
+    int create(void)
     {
-        return false;
-    }
+        // base initialization
+        D3DSample::create();
 
-    D3DPRESENT_PARAMETERS d3dpp;
-    ZeroMemory(&d3dpp,sizeof(D3DPRESENT_PARAMETERS));
+        // initialize DirectX
+        HRESULT r;
 
-    DWORD flags = D3DCREATE_HARDWARE_VERTEXPROCESSING | D3DCREATE_PUREDEVICE | D3DCREATE_NOWINDOWCHANGES
-            | D3DCREATE_MULTITHREADED;
-
-    d3dpp.Windowed = true;
-    d3dpp.Flags = 0;
-    d3dpp.BackBufferCount = 0;
-    d3dpp.BackBufferFormat = D3DFMT_A8R8G8B8;
-    d3dpp.BackBufferHeight = HEIGHT;
-    d3dpp.BackBufferWidth = WIDTH;
-    d3dpp.MultiSampleType = D3DMULTISAMPLE_NONE;
-    d3dpp.SwapEffect = D3DSWAPEFFECT_DISCARD;
-    d3dpp.hDeviceWindow = hWnd;
-    d3dpp.PresentationInterval = D3DPRESENT_INTERVAL_IMMEDIATE;
-    d3dpp.FullScreen_RefreshRateInHz = D3DPRESENT_RATE_DEFAULT;
-
-    if (FAILED(pD3D->CreateDevice(D3DADAPTER_DEFAULT, D3DDEVTYPE_HAL, hWnd, flags, &d3dpp, &dev)))
-    {
-        return false;
-    }
-
-    if (FAILED(dev->GetBackBuffer(0, 0, D3DBACKBUFFER_TYPE_MONO, &pBackBuffer)))
-    {
-        return false;
-    }
-
-    return true;
-}
-
-bool initDirect3DTextures()
-{
-    // Note: sharing is not supported on some platforms
-    if (FAILED(dev->CreateOffscreenPlainSurface(WIDTH, HEIGHT, D3DFMT_A8R8G8B8, D3DPOOL_DEFAULT, &pSurface, NULL/*&surfaceShared*/)))
-    {
-        std::cerr << "Can't create surface for result" << std::endl;
-        return false;
-    }
-
-    // Note: sharing is not supported on some platforms
-    if (FAILED(dev->CreateOffscreenPlainSurface(WIDTH, HEIGHT, D3DFMT_A8R8G8B8, D3DPOOL_DEFAULT, &pReadOnlySurface, NULL/*&readOnlySurfaceShared*/)))
-    {
-        std::cerr << "Can't create read only surface" << std::endl;
-        return false;
-    }
-    else
-    {
-        IDirect3DSurface9* pTmpSurface;
-        if (FAILED(dev->CreateOffscreenPlainSurface(WIDTH, HEIGHT, D3DFMT_A8R8G8B8, D3DPOOL_DEFAULT, &pTmpSurface, NULL)))
+        m_pD3D9 = ::Direct3DCreate9(D3D_SDK_VERSION);
+        if (NULL == m_pD3D9)
         {
-            std::cerr << "Can't create temp surface for CPU write" << std::endl;
-            return false;
+            return -1;
         }
 
-        D3DLOCKED_RECT memDesc = {0, NULL};
-        RECT rc = {0, 0, WIDTH, HEIGHT};
-        if (SUCCEEDED(pTmpSurface->LockRect(&memDesc, &rc, 0)))
+        DWORD flags = D3DCREATE_HARDWARE_VERTEXPROCESSING |
+                      D3DCREATE_PUREDEVICE |
+                      D3DCREATE_NOWINDOWCHANGES |
+                      D3DCREATE_MULTITHREADED |
+                      D3DCREATE_FPU_PRESERVE;
+
+        D3DPRESENT_PARAMETERS d3dpp;
+        ::ZeroMemory(&d3dpp, sizeof(D3DPRESENT_PARAMETERS));
+
+        d3dpp.Windowed                   = true;
+        d3dpp.Flags                      = 0;
+        d3dpp.BackBufferCount            = 0;
+        d3dpp.BackBufferFormat           = D3DFMT_A8R8G8B8;
+        d3dpp.BackBufferHeight           = m_height;
+        d3dpp.BackBufferWidth            = m_width;
+        d3dpp.MultiSampleType            = D3DMULTISAMPLE_NONE;
+        d3dpp.SwapEffect                 = D3DSWAPEFFECT_DISCARD;
+        d3dpp.hDeviceWindow              = m_hWnd;
+        d3dpp.PresentationInterval       = D3DPRESENT_INTERVAL_IMMEDIATE;
+        d3dpp.FullScreen_RefreshRateInHz = D3DPRESENT_RATE_DEFAULT;
+
+        r = m_pD3D9->CreateDevice(D3DADAPTER_DEFAULT, D3DDEVTYPE_HAL, m_hWnd, flags, &d3dpp, &m_pD3D9Dev);
+        if (FAILED(r))
         {
-            cv::Mat m(cv::Size(WIDTH, HEIGHT), CV_8UC4, memDesc.pBits, (int)memDesc.Pitch);
-            getInputTexture().copyTo(m);
-            pTmpSurface->UnlockRect();
-            dev->StretchRect(pTmpSurface, NULL, pReadOnlySurface, NULL, D3DTEXF_NONE);
+            return -1;
         }
-        else
+
+        r = m_pD3D9Dev->GetBackBuffer(0, 0, D3DBACKBUFFER_TYPE_MONO, &m_pBackBuffer);
+        if (FAILED(r))
         {
-            std::cerr << "Can't LockRect() on surface" << std::endl;
+            return -1;
         }
-        pTmpSurface->Release();
-    }
 
-    if (FAILED(dev->CreateOffscreenPlainSurface(WIDTH, HEIGHT, D3DFMT_A8R8G8B8, D3DPOOL_DEFAULT, &pCPUWriteSurface, NULL)))
-    {
-        std::cerr << "Can't create surface for CPU write" << std::endl;
-        return false;
-    }
-
-    return true;
-}
-
-void render(void)
-{
-    // check to make sure you have a valid Direct3D device
-    CV_Assert(dev);
-
-    renderToD3DObject();
-
-    if (g_sampleType == 0)
-    {
-        // nothing
-    }
-    else if (g_sampleType == 1)
-    {
-        if (FAILED(dev->StretchRect(pCPUWriteSurface, NULL, pBackBuffer, NULL, D3DTEXF_NONE)))
+        r = m_pD3D9Dev->CreateOffscreenPlainSurface(m_width, m_height, D3DFMT_A8R8G8B8, D3DPOOL_DEFAULT, &m_pSurface, NULL);
+        if (FAILED(r))
         {
-            std::cerr << "Can't StretchRect()" << std::endl;
+            std::cerr << "Can't create surface for result" << std::endl;
+            return -1;
         }
-    }
-    else
-    {
-        if (FAILED(dev->StretchRect(pSurface, NULL, pBackBuffer, NULL, D3DTEXF_NONE)))
+
+        // initialize OpenCL context of OpenCV lib from DirectX
+        if (cv::ocl::haveOpenCL())
         {
-            std::cerr << "Can't StretchRect()" << std::endl;
+            m_oclCtx = cv::directx::ocl::initializeContextFromDirect3DDevice9(m_pD3D9Dev);
         }
-    }
 
-    if (SUCCEEDED(dev -> BeginScene()))
+        m_oclDevName = cv::ocl::useOpenCL() ?
+            cv::ocl::Context::getDefault().device(0).name() :
+            "No OpenCL device";
+
+        return 0;
+    } // create()
+
+
+    // get media data on DX surface for further processing
+    int get_surface(LPDIRECT3DSURFACE9* ppSurface)
     {
-        // end the scene
-        dev -> EndScene();
-    }
+        HRESULT r;
 
-    // present the back buffer contents to the display
-    dev->Present(NULL, NULL, NULL, NULL);
-}
+        if (!m_cap.read(m_frame_bgr))
+            return -1;
 
-void cleanUp (void)
-{
-    SAFE_RELEASE(pCPUWriteSurface);
-    SAFE_RELEASE(pReadOnlySurface);
-    SAFE_RELEASE(pSurface);
-    SAFE_RELEASE(pBackBuffer);
-    SAFE_RELEASE(dev);
-    SAFE_RELEASE(pD3D);}
+        cv::cvtColor(m_frame_bgr, m_frame_rgba, CV_RGB2RGBA);
+
+        D3DLOCKED_RECT memDesc = { 0, NULL };
+        RECT rc = { 0, 0, m_width, m_height };
+
+        r = m_pSurface->LockRect(&memDesc, &rc, 0);
+        if (FAILED(r))
+        {
+            return r;
+        }
+
+        cv::Mat m(m_height, m_width, CV_8UC4, memDesc.pBits, memDesc.Pitch);
+        // copy video frame data to surface
+        m_frame_rgba.copyTo(m);
+
+        r = m_pSurface->UnlockRect();
+        if (FAILED(r))
+        {
+            return r;
+        }
+
+        *ppSurface = m_pSurface;
+
+        return 0;
+    } // get_surface()
+
+
+    // process and render media data
+    int render()
+    {
+        try
+        {
+            if (m_shutdown)
+                return 0;
+
+            HRESULT r;
+            LPDIRECT3DSURFACE9 pSurface;
+
+            r = get_surface(&pSurface);
+            if (FAILED(r))
+            {
+                return -1;
+            }
+
+            switch (m_mode)
+            {
+                case MODE_NOP:
+                    // no processing
+                    break;
+
+                case MODE_CPU:
+                {
+                    // process video frame on CPU
+                    D3DLOCKED_RECT memDesc = { 0, NULL };
+                    RECT rc = { 0, 0, m_width, m_height };
+
+                    r = pSurface->LockRect(&memDesc, &rc, 0);
+                    if (FAILED(r))
+                    {
+                        return -1;
+                    }
+
+                    cv::Mat m(m_height, m_width, CV_8UC4, memDesc.pBits, memDesc.Pitch);
+
+                    if (!m_disableProcessing)
+                    {
+                        // blur D3D9 surface with OpenCV on CPU
+                        cv::blur(m, m, cv::Size(15, 15), cv::Point(-7, -7));
+                    }
+
+                    r = pSurface->UnlockRect();
+                    if (FAILED(r))
+                    {
+                        return -1;
+                    }
+
+                    break;
+                }
+
+                case MODE_GPU:
+                {
+                    // process video frame on GPU
+                    cv::UMat u;
+
+                    cv::directx::convertFromDirect3DSurface9(pSurface, u);
+
+                    if (!m_disableProcessing)
+                    {
+                        // blur D3D9 surface with OpenCV on GPU with OpenCL
+                        cv::blur(u, u, cv::Size(15, 15), cv::Point(-7, -7));
+                    }
+
+                    cv::directx::convertToDirect3DSurface9(u, pSurface);
+
+                    break;
+                }
+
+            } // switch
+
+            print_info(pSurface, m_mode, getFps(), m_oclDevName);
+
+            // traditional DX render pipeline:
+            //   BitBlt surface to backBuffer and flip backBuffer to frontBuffer
+            r = m_pD3D9Dev->StretchRect(pSurface, NULL, m_pBackBuffer, NULL, D3DTEXF_NONE);
+            if (FAILED(r))
+            {
+                return -1;
+            }
+
+            // present the back buffer contents to the display
+            r = m_pD3D9Dev->Present(NULL, NULL, NULL, NULL);
+            if (FAILED(r))
+            {
+                return -1;
+            }
+        }  // try
+
+        catch (cv::Exception& e)
+        {
+            std::cerr << "Exception: " << e.what() << std::endl;
+            return 10;
+        }
+
+        return 0;
+    } // render()
+
+
+    void print_info(LPDIRECT3DSURFACE9 pSurface, int mode, float fps, cv::String oclDevName)
+    {
+        HDC hDC;
+
+        HRESULT r = pSurface->GetDC(&hDC);
+        if (FAILED(r))
+        {
+            return;
+        }
+
+        HFONT hFont = (HFONT)::GetStockObject(SYSTEM_FONT);
+
+        HFONT hOldFont = (HFONT)::SelectObject(hDC, hFont);
+
+        if (hOldFont)
+        {
+            TEXTMETRIC tm;
+            ::GetTextMetrics(hDC, &tm);
+
+            char buf[256];
+            int  y = 0;
+
+            buf[0] = 0;
+            sprintf(buf, "Mode: %s", m_modeStr[mode].c_str());
+            ::TextOut(hDC, 0, y, buf, (int)strlen(buf));
+
+            y += tm.tmHeight;
+            buf[0] = 0;
+            sprintf(buf, "FPS: %2.1f", fps);
+            ::TextOut(hDC, 0, y, buf, (int)strlen(buf));
+
+            y += tm.tmHeight;
+            buf[0] = 0;
+            sprintf(buf, "OpenCL device: %s", oclDevName.c_str());
+            ::TextOut(hDC, 0, y, buf, (int)strlen(buf));
+
+            ::SelectObject(hDC, hOldFont);
+        }
+
+        r = pSurface->ReleaseDC(hDC);
+
+        return;
+    } // print_info()
+
+
+    int cleanup(void)
+    {
+        SAFE_RELEASE(m_pSurface);
+        SAFE_RELEASE(m_pBackBuffer);
+        SAFE_RELEASE(m_pD3D9Dev);
+        SAFE_RELEASE(m_pD3D9);
+        D3DSample::cleanup();
+        return 0;
+    } // cleanup()
+
+private:
+    LPDIRECT3D9        m_pD3D9;
+    LPDIRECT3DDEVICE9  m_pD3D9Dev;
+    LPDIRECT3DSURFACE9 m_pBackBuffer;
+    LPDIRECT3DSURFACE9 m_pSurface;
+    cv::ocl::Context   m_oclCtx;
+    cv::String         m_oclPlatformName;
+    cv::String         m_oclDevName;
+};
+
+
+// main func
+ENTRY_POINT(D3D9WinApp, "D3D9 interop sample");
