--- samples/directx/d3d9ex_opencl.cpp
@@ -1,396 +1,0 @@
-#define WIN32_LEAN_AND_MEAN
-#include <windows.h>
-#include <d3d9.h>
-
-#include <string>
-#include <iostream>
-#include <queue>
-
-#include "opencv2/core.hpp"
-#include "opencv2/core/directx.hpp"
-#include "opencv2/core/ocl.hpp"
-#include "opencv2/imgproc.hpp"
-#include "opencv2/videoio.hpp"
-#include "winapp.hpp"
-
-#pragma comment (lib, "d3d9.lib")
-
-
-class D3D9ExWinApp : public WinApp
-{
-public:
-    D3D9ExWinApp(int width, int height, std::string& window_name, cv::VideoCapture& cap) :
-        WinApp(width, height, window_name)
-    {
-        m_shutdown = false;
-        m_mode = 0;
-        m_modeStr[0] = cv::String("No processing");
-        m_modeStr[1] = cv::String("Processing on CPU");
-        m_modeStr[2] = cv::String("Processing on GPU");
-        m_disableProcessing = false;
-        m_cap = cap;
-    }
-
-    ~D3D9ExWinApp() {}
-
-    int onClose(void)
-    {
-        m_shutdown = true;
-        cleanup();
-        ::DestroyWindow(m_hWnd);
-        return 0;
-    }
-
-    virtual LRESULT CALLBACK WndProc(HWND hWnd, UINT message, WPARAM wParam, LPARAM lParam)
-    {
-        switch (message)
-        {
-        case WM_CHAR:
-            if (wParam >= '0' && wParam <= '2')
-            {
-                m_mode = (char)wParam - '0';
-                return 0;
-            }
-            else if (wParam == VK_SPACE)
-            {
-                m_disableProcessing = !m_disableProcessing;
-                return 0;
-            }
-            else if (wParam == VK_ESCAPE)
-            {
-                return onClose();
-            }
-            break;
-
-        case WM_CLOSE:
-            return onClose();
-
-        case WM_DESTROY:
-            ::PostQuitMessage(0);
-            return 0;
-        }
-
-        return ::DefWindowProc(hWnd, message, wParam, lParam);
-    }
-
-    static float getFps()
-    {
-        static std::queue<int64> time_queue;
-
-        int64 now = cv::getTickCount();
-        int64 then = 0;
-        time_queue.push(now);
-
-        if (time_queue.size() >= 2)
-            then = time_queue.front();
-
-        if (time_queue.size() >= 25)
-            time_queue.pop();
-
-        return time_queue.size() * (float)cv::getTickFrequency() / (now - then);
-    }
-
-    int init(void)
-    {
-        HRESULT r;
-
-        r = ::Direct3DCreate9Ex(D3D_SDK_VERSION, &m_pD3D9Ex);
-        if (FAILED(r))
-        {
-            return -1;
-        }
-
-        DWORD flags = D3DCREATE_HARDWARE_VERTEXPROCESSING |
-            D3DCREATE_PUREDEVICE |
-            D3DCREATE_NOWINDOWCHANGES |
-            D3DCREATE_MULTITHREADED |
-            D3DCREATE_FPU_PRESERVE;
-
-        D3DPRESENT_PARAMETERS d3dpp;
-        ::ZeroMemory(&d3dpp, sizeof(D3DPRESENT_PARAMETERS));
-
-        d3dpp.Windowed                   = true;
-        d3dpp.Flags                      = 0;
-        d3dpp.BackBufferCount            = 0;
-        d3dpp.BackBufferFormat           = D3DFMT_A8R8G8B8;
-        d3dpp.BackBufferHeight           = m_height;
-        d3dpp.BackBufferWidth            = m_width;
-        d3dpp.MultiSampleType            = D3DMULTISAMPLE_NONE;
-        d3dpp.SwapEffect                 = D3DSWAPEFFECT_DISCARD;
-        d3dpp.hDeviceWindow              = m_hWnd;
-        d3dpp.PresentationInterval       = D3DPRESENT_INTERVAL_IMMEDIATE;
-        d3dpp.FullScreen_RefreshRateInHz = D3DPRESENT_RATE_DEFAULT;
-
-        r = m_pD3D9Ex->CreateDeviceEx(D3DADAPTER_DEFAULT, D3DDEVTYPE_HAL, m_hWnd, flags, &d3dpp, NULL, &m_pD3D9DevEx);
-        if (FAILED(r))
-        {
-            return -1;
-        }
-
-        r = m_pD3D9DevEx->GetBackBuffer(0, 0, D3DBACKBUFFER_TYPE_MONO, &m_pBackBuffer);
-        if (FAILED(r))
-        {
-            return -1;
-        }
-
-        r = m_pD3D9DevEx->CreateOffscreenPlainSurface(m_width, m_height, D3DFMT_A8R8G8B8, D3DPOOL_DEFAULT, &m_pSurface, NULL);
-        if (FAILED(r))
-        {
-            std::cerr << "Can't create surface for result" << std::endl;
-            return -1;
-        }
-
-        if (cv::ocl::haveOpenCL())
-        {
-            m_oclCtx = cv::directx::ocl::initializeContextFromDirect3DDevice9(m_pD3D9DevEx);
-        }
-
-        m_oclDevName = cv::ocl::useOpenCL() ?
-            cv::ocl::Context::getDefault().device(0).name() :
-            "No OpenCL device";
-
-        return 0;
-    } // init()
-
-
-    int get_surface(LPDIRECT3DSURFACE9* ppSurface)
-    {
-        HRESULT r;
-
-        if (!m_cap.read(m_frame_bgr))
-            return -1;
-
-        cv::cvtColor(m_frame_bgr, m_frame_rgba, CV_RGB2RGBA);
-
-        D3DLOCKED_RECT memDesc = { 0, NULL };
-        RECT rc = { 0, 0, m_width, m_height };
-
-        r = m_pSurface->LockRect(&memDesc, &rc, 0);
-        if (FAILED(r))
-        {
-            return r;
-        }
-
-        cv::Mat m(m_height, m_width, CV_8UC4, memDesc.pBits, memDesc.Pitch);
-        // copy video frame data to surface
-        m_frame_rgba.copyTo(m);
-
-        r = m_pSurface->UnlockRect();
-        if (FAILED(r))
-        {
-            return r;
-        }
-
-        *ppSurface = m_pSurface;
-
-        return 0;
-    }
-
-
-    void print_info(LPDIRECT3DSURFACE9 pSurface, int mode, float fps, cv::String oclDevName)
-    {
-        HDC hDC;
-
-        HRESULT r = pSurface->GetDC(&hDC);
-        if (FAILED(r))
-        {
-            return;
-        }
-
-        HFONT hFont = (HFONT)::GetStockObject(SYSTEM_FONT);
-
-        HFONT hOldFont = (HFONT)::SelectObject(hDC, hFont);
-
-        if (hOldFont)
-        {
-            TEXTMETRIC tm;
-            ::GetTextMetrics(hDC, &tm);
-
-            char buf[256];
-            int  y = 0;
-
-            buf[0] = 0;
-            sprintf(buf, "Mode: %s", m_modeStr[mode].c_str());
-            ::TextOut(hDC, 0, y, buf, (int)strlen(buf));
-
-            y += tm.tmHeight;
-            buf[0] = 0;
-            sprintf(buf, "FPS: %2.1f", fps);
-            ::TextOut(hDC, 0, y, buf, (int)strlen(buf));
-
-            y += tm.tmHeight;
-            buf[0] = 0;
-            sprintf(buf, "OpenCL device: %s", oclDevName.c_str());
-            ::TextOut(hDC, 0, y, buf, (int)strlen(buf));
-
-            ::SelectObject(hDC, hOldFont);
-        }
-
-        r = pSurface->ReleaseDC(hDC);
-
-        return;
-    }
-
-
-    int render()
-    {
-        try
-        {
-            if (m_shutdown)
-                return 0;
-
-            HRESULT r;
-            LPDIRECT3DSURFACE9 pSurface;
-
-            r = get_surface(&pSurface);
-            if (FAILED(r))
-            {
-                return -1;
-            }
-
-            switch (m_mode)
-            {
-                case 0:
-                    // no processing
-                    break;
-
-                case 1:
-                {
-                    // process video frame on CPU
-                    D3DLOCKED_RECT memDesc = { 0, NULL };
-                    RECT rc = { 0, 0, m_width, m_height };
-
-                    r = pSurface->LockRect(&memDesc, &rc, 0);
-                    if (FAILED(r))
-                    {
-                        return -1;
-                    }
-
-                    cv::Mat m(m_height, m_width, CV_8UC4, memDesc.pBits, memDesc.Pitch);
-
-                    if (!m_disableProcessing)
-                    {
-                        // blur D3D9 surface with OpenCV on CPU
-                        cv::blur(m, m, cv::Size(15, 15), cv::Point(-7, -7));
-                    }
-
-                    r = pSurface->UnlockRect();
-                    if (FAILED(r))
-                    {
-                        return -1;
-                    }
-
-                    break;
-                }
-
-                case 2:
-                {
-                    // process video frame on GPU
-                    cv::UMat u;
-
-                    cv::directx::convertFromDirect3DSurface9(pSurface, u);
-
-                    if (!m_disableProcessing)
-                    {
-                        // blur D3D9 surface with OpenCV on GPU with OpenCL
-                        cv::blur(u, u, cv::Size(15, 15), cv::Point(-7, -7));
-                    }
-
-                    cv::directx::convertToDirect3DSurface9(u, pSurface);
-
-                    break;
-                }
-
-            } // switch
-
-            print_info(pSurface, m_mode, getFps(), m_oclDevName);
-
-            // traditional DX render pipeline:
-            //   BitBlt surface to backBuffer and flip backBuffer to frontBuffer
-            r = m_pD3D9DevEx->StretchRect(pSurface, NULL, m_pBackBuffer, NULL, D3DTEXF_NONE);
-            if (FAILED(r))
-            {
-                return -1;
-            }
-
-            // present the back buffer contents to the display
-            r = m_pD3D9DevEx->Present(NULL, NULL, NULL, NULL);
-            if (FAILED(r))
-            {
-                return -1;
-            }
-        }
-
-        catch (cv::Exception& e)
-        {
-            std::cerr << "Exception: " << e.what() << std::endl;
-            return 10;
-        }
-
-        return 0;
-    }
-
-    int cleanup(void)
-    {
-        SAFE_RELEASE(m_pSurface);
-        SAFE_RELEASE(m_pBackBuffer);
-        SAFE_RELEASE(m_pD3D9DevEx);
-        SAFE_RELEASE(m_pD3D9Ex);
-        return 0;
-    }
-
-private:
-    bool                 m_shutdown;
-    int                  m_mode;
-    cv::String           m_modeStr[3];
-    int                  m_disableProcessing;
-    LPDIRECT3D9EX        m_pD3D9Ex;
-    LPDIRECT3DDEVICE9EX  m_pD3D9DevEx;
-    LPDIRECT3DSURFACE9   m_pBackBuffer;
-    LPDIRECT3DSURFACE9   m_pSurface;
-    cv::VideoCapture     m_cap;
-    cv::Mat              m_frame_bgr;
-    cv::Mat              m_frame_rgba;
-    cv::ocl::Context     m_oclCtx;
-    cv::String           m_oclPlatformName;
-    cv::String           m_oclDevName;
-};
-
-
-using namespace cv;
-
-int main(int argc, char** argv)
-{
-    cv::VideoCapture cap;
-
-    if (argc > 1)
-    {
-        cap.open(argv[1]);
-    }
-    else
-        cap.open(0);
-
-    int width = (int)cap.get(CAP_PROP_FRAME_WIDTH);
-    int height = (int)cap.get(CAP_PROP_FRAME_HEIGHT);
-    std::string wndname = "D3D9Ex Window";
-
-    D3D9ExWinApp app(width, height, wndname, cap);
-
-    try
-    {
-        app.Create();
-        return app.run();
-    }
-
-    catch (cv::Exception& e)
-    {
-        std::cerr << "Exception: " << e.what() << std::endl;
-        return 10;
-    }
-
-    catch (...)
-    {
-        std::cerr << "FATAL ERROR: Unknown exception" << std::endl;
-        return 11;
-    }
-}
